requirements.txt
requests>=2.31.0
pytrends>=4.9.0
openai>=1.3.0
google-api-python-client>=2.108.0
isodate>=0.6.1
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
python-dotenv>=1.0.0
aiohttp>=3.9.1
asyncio>=3.4.3
typing-extensions>=4.8.0
python-json-logger>=2.0.7
beautifulsoup4>=4.12.2
selenium>=4.15.2
webdriver-manager>=4.0.1
lxml>=4.9.3
pytest>=8.0.0
pytest-asyncio>=0.25.0 